+++ R/fct_spectra_curation.R
@@ -494,14 +494,16 @@ calculate_cut_offs <- function(summarized_checks,
     dplyr::reframe(
       cut_off_sum_intensity = if (use_mean_SD) { 
         mean_plus_SD(sum_intensity, SD_factor, uncalibrated_as_NA) } else { 
-          quantile(sum_intensity, 
-                   probs = percentile / 100,
-                   names = FALSE,
-                   na.rm = uncalibrated_as_NA) },
-      cut_off_passing_analyte_percentage = quantile(passing_analyte_percentage, 
-                              probs = percentile / 100,
-                              names = FALSE,
-                              na.rm = uncalibrated_as_NA),
+          percentile_cutoff(
+            data = sum_intensity,
+            percentage_to_exclude = percentile,
+            remove_NA = uncalibrated_as_NA
+          )},
+      cut_off_passing_analyte_percentage = percentile_cutoff(
+        data = passing_analyte_percentage,
+        percentage_to_exclude = percentile,
+        remove_NA = uncalibrated_as_NA
+      ),
       sample_type = unique(sample_type),
       curation_method = ifelse(is.null(control_sample_types), 
                                "based_on_percentiles",
@@ -579,8 +581,8 @@ curate_spectra <- function(checked_data, summarized_checks, cut_offs) {
   
   curated_spectra <- summarized_checks_with_cut_offs %>% 
     # Can't use all() instead of & because all() is not vectorized
-    dplyr::mutate(has_passed_spectra_curation = passing_analyte_percentage > cut_off_passing_analyte_percentage &
-                    sum_intensity > cut_off_sum_intensity,
+    dplyr::mutate(has_passed_spectra_curation = passing_analyte_percentage >= cut_off_passing_analyte_percentage &
+                    sum_intensity >= cut_off_sum_intensity,
                   has_passed_spectra_curation = ifelse(uncalibrated, 
                                                        FALSE,
                                                        has_passed_spectra_curation)) %>%